fix: close managed markdown review gaps
